Stuff (read description)
- Solution-wide reformat + cleanup.
- Rename AppEngine to WebServer.
- Create a basic networking lib for inter-server communication (dropping Redis soon)
- Moved some stuff around.
- Renamed DatabaseManager to DbServer. Will be using it as the db operation handler.

+++ RealmServer/GameServer/Game/Entities/Player/Player.Chat.cs
@@ -20,7 +20,7 @@ public partial class Player
             "",
             0,
             -1,
-            (byte)0,
+            0,
             null,
             text));
     }
@@ -39,10 +39,10 @@ public partial class Player
     public void SendError(string text)
     {
         User.SendPacket(new Text(
-                    "*Error*",
+            "*Error*",
             0,
             -1,
-            (byte)0,
+            0,
             null,
             text));
     }
@@ -50,22 +50,22 @@ public partial class Player
     public void SendHelp(string text)
     {
         User.SendPacket(new Text(
-                    "*Help*",
+            "*Help*",
             0,
             -1,
-            (byte)0,
+            0,
             null,
             text));
     }
 
     public void SendEnemy(Entity entity, string text)
     {
-        User.SendPacket(new Text($"#{entity.Desc.DisplayName}", entity.Id, -1, (byte)3, null, text));
+        User.SendPacket(new Text($"#{entity.Desc.DisplayName}", entity.Id, -1, 3, null, text));
     }
 
     public void SendEnemy(string name, string text)
     {
-        User.SendPacket(new Text($"#{name}", -1, -1, (byte)3, null, text));
+        User.SendPacket(new Text($"#{name}", -1, -1, 3, null, text));
     }
 
     public void Speak(string text)
@@ -90,7 +90,7 @@ public partial class Player
         {
             var user = plr.User;
             if (!user.Account.IgnoredIds?.Contains(acc.AccountId) ?? true)
-                User.SendPacket(new Text(acc.Admin ? $"@{acc.Name}" : acc.Name, Id, NumStars, (byte)5, null, text));
+                User.SendPacket(new Text(acc.Admin ? $"@{acc.Name}" : acc.Name, Id, NumStars, 5, null, text));
         });
     }
